Simplify diameter/offset computation logic
There is no need to enforce a stepping based on 8 pixels. Instead,
compute the hole offset first as 1/8th of the diameter, and derive
everything else from there. To ensure symmetry, it's sufficient to
round down the hole offset to an even number when there is a ring
(i.e. if hole_offset > 1).